Tasty Reader: unpickle sealed children more lazily

+++ src/compiler/scala/tools/nsc/tasty/bridge/ContextOps.scala
@@ -13,12 +13,17 @@
 package scala.tools.nsc.tasty.bridge
 
 import scala.annotation.tailrec
+
+import scala.collection.mutable
 import scala.reflect.io.AbstractFile
+import scala.reflect.internal.MissingRequirementError
 
 import scala.tools.tasty.{TastyName, TastyFlags}, TastyFlags._, TastyName.ObjectName
 import scala.tools.nsc.tasty.{TastyUniverse, TastyModes, SafeEq}, TastyModes._
-import scala.reflect.internal.MissingRequirementError
-import scala.collection.mutable
+import scala.tools.nsc.tasty.{cyan, yellow, magenta, blue, green}
+
+import scala.util.chaining._
+
 
 /**This contains the definition for `Context`, along with standard error throwing capabilities with user friendly
  * formatted errors that can change their output depending on the context mode.
@@ -111,15 +116,45 @@ trait ContextOps { self: TastyUniverse =>
    * sealed child.
    */
   private def analyseAnnotations(sym: Symbol)(implicit ctx: Context): Unit = {
+
+    def lookupChild(childTpe: Type): Symbol = {
+      val child = symOfType(childTpe)
+      assert(isSymbol(child), s"did not find symbol of sealed child ${showType(childTpe)}")
+      if (child.isClass) {
+        child
+      }
+      else {
+        assert(child.isModule, s"sealed child was not class or object ${showSym(child)}")
+        child.moduleClass
+      }
+    }
+
     for (annot <- sym.annotations) {
       annot.completeInfo()
       if (annot.tpe.typeSymbolDirect === defn.ChildAnnot) {
-        val child = annot.tpe.typeArgs.head.typeSymbolDirect
-        sym.addChild(child)
+        val child = {
+          val child0 = lookupChild(annot.tpe.typeArgs.head)
+          if (child0 eq sym) {
+            // dotty represents a local sealed child of `C` with a child annotation
+            // that directly references `C`, this causes an infinite loop in
+            // `sealedDescendants`. See the tests:
+            // - test/tasty/neg/src-3/dottyi3149/dotty_i3149.scala
+            // - test/tasty/neg/src-2/Testdotty_i3149_fail.scala
+            // TODO [tasty] - fix assumption in compiler that sealed children cannot
+            // contain the parent class
+            ctx.newLocalSealedChildProxy(sym)
+          }
+          else {
+            child0
+          }
+        }
         ctx.log(s"adding sealed child ${showSym(child)} to ${showSym(sym)}")
+        sym.addChild(child)
       }
     }
   }
+
+  final case class TraceInfo[-T](query: String, qual: String, res: T => String, modifiers: List[String] = Nil)
 
   /**Maintains state through traversal of a TASTy file, such as the outer scope of the defintion being traversed, the
    * traversal mode, and the root owners and source path for the TASTy file.
@@ -155,11 +190,42 @@ trait ContextOps { self: TastyUniverse =>
     }
 
     final def log(str: => String): Unit = {
-      if (u.settings.YdebugTasty)
-        u.reporter.echo(
-          pos = u.NoPosition,
-          msg = str.linesIterator.map(line => s"#[$classRoot]: $line").mkString(System.lineSeparator)
+      if (u.settings.YdebugTasty) {
+        logImpl(str)
+      }
+    }
+
+    private final def logImpl(str: => String): Unit = u.reporter.echo(
+      pos = u.NoPosition,
+      msg = str
+              .linesIterator
+              .map(line => s"${blue(s"${showSymStable(classRoot)}:")} $line")
+              .mkString(System.lineSeparator)
+    )
+
+    @inline final def trace[T](info: => TraceInfo[T])(op: => T): T = {
+
+      def withTrace(info: => TraceInfo[T], op: => T)(traceId: String): T = {
+        val i = info
+        val modStr = (
+          if (i.modifiers.isEmpty) ""
+          else " " + green(i.modifiers.mkString("[", ",", "]"))
         )
+        logImpl(s"${yellow(s"$traceId")} ${cyan(s"<<< ${i.query}:")} ${magenta(i.qual)}$modStr")
+        op.tap(eval => logImpl(s"${yellow(s"$traceId")} ${cyan(s">>>")} ${magenta(i.res(eval))}$modStr"))
+      }
+
+      if (u.settings.YdebugTasty) initialContext.addFrame(withTrace(info, op))
+      else op
+    }
+
+    /** Trace only when `-Vdebug` is set
+     */
+    @inline final def traceV[T](info: => TraceInfo[T])(op: => T): T = {
+      if (u.settings.debug.value) {
+        trace(info)(op)
+      }
+      else op
     }
 
     def owner: Symbol
@@ -204,6 +270,17 @@ trait ContextOps { self: TastyUniverse =>
       flags = Method,
       info  = info
     )
+
+    final def newLocalSealedChildProxy(cls: Symbol): Symbol = {
+      val tflags = Private | Local
+      unsafeNewClassSymbol(
+        owner = cls,
+        typeName = TastyName.SimpleName(cls.fullName('$') + "$$localSealedChildProxy").toTypeName,
+        flags = tflags,
+        info = defn.LocalSealedChildProxyInfo(cls, tflags),
+        privateWithin = u.NoSymbol
+      )
+    }
 
     final def findRootSymbol(roots: Set[Symbol], name: TastyName): Option[Symbol] = {
       import TastyName.TypeName
@@ -256,11 +333,11 @@ trait ContextOps { self: TastyUniverse =>
     /** Guards the creation of an object val by checking for an existing definition in the owner's scope
       */
     final def delayCompletion(owner: Symbol, name: TastyName, completer: TastyCompleter, privateWithin: Symbol = noSymbol): Symbol = {
-      def default() = unsafeNewSymbol(owner, name, completer.originalFlagSet, completer, privateWithin)
-      if (completer.originalFlagSet.is(Object)) {
+      def default() = unsafeNewSymbol(owner, name, completer.tflags, completer, privateWithin)
+      if (completer.tflags.is(Object)) {
         val sourceObject = findObject(owner, encodeTermName(name))
         if (isSymbol(sourceObject))
-          redefineSymbol(sourceObject, completer.originalFlagSet, completer, privateWithin)
+          redefineSymbol(sourceObject, completer.tflags, completer, privateWithin)
         else
           default()
       }
@@ -272,11 +349,11 @@ trait ContextOps { self: TastyUniverse =>
     /** Guards the creation of an object class by checking for an existing definition in the owner's scope
       */
     final def delayClassCompletion(owner: Symbol, typeName: TastyName.TypeName, completer: TastyCompleter, privateWithin: Symbol): Symbol = {
-      def default() = unsafeNewClassSymbol(owner, typeName, completer.originalFlagSet, completer, privateWithin)
-      if (completer.originalFlagSet.is(Object)) {
+      def default() = unsafeNewClassSymbol(owner, typeName, completer.tflags, completer, privateWithin)
+      if (completer.tflags.is(Object)) {
         val sourceObject = findObject(owner, encodeTermName(typeName.toTermName))
         if (isSymbol(sourceObject))
-          redefineSymbol(sourceObject.objectImplementation, completer.originalFlagSet, completer, privateWithin)
+          redefineSymbol(sourceObject.objectImplementation, completer.tflags, completer, privateWithin)
         else
           default()
       }
@@ -286,8 +363,10 @@ trait ContextOps { self: TastyUniverse =>
     }
 
     def evict(sym: Symbol): Unit = {
-      sym.owner.rawInfo.decls.unlink(sym)
-      sym.info = u.NoType
+      if (isSymbol(sym)) {
+        sym.owner.rawInfo.decls.unlink(sym)
+        sym.info = u.NoType
+      }
     }
 
     final def enterIfUnseen(sym: Symbol): Unit = {
@@ -340,9 +419,14 @@ trait ContextOps { self: TastyUniverse =>
         }
       }
       else if (flags.is(FlagSets.Creation.ObjectDef)) {
-        log(s"!!! visited module value $name first")
+        val isEnum = flags.is(FlagSets.SingletonEnum)
+        if (!isEnum) {
+          log(s"!!! visited module value $name first")
+        }
         val module = owner.newModule(encodeTermName(name), u.NoPosition, newSymbolFlagSet(flags))
-        module.moduleClass.info = defn.DefaultInfo
+        module.moduleClass.info =
+          if (isEnum) defn.SingletonEnumClassInfo(module, flags)
+          else defn.DefaultInfo
         module
       }
       else if (name.isTypeName) {
@@ -374,20 +458,20 @@ trait ContextOps { self: TastyUniverse =>
 
     final def enterClassCompletion(): Symbol = {
       val cls = globallyVisibleOwner.asClass
-      val assumedSelfType =
-        if (cls.is(Object) && cls.owner.isClass) defn.SingleType(cls.owner.thisType, cls.sourceModule)
-        else u.NoType
-      cls.info = u.ClassInfoType(cls.repr.parents, cls.repr.decls, assumedSelfType.typeSymbolDirect)
+      val assumedSelfSym = {
+        if (cls.is(Object) && cls.owner.isClass) {
+          cls.sourceModule
+        }
+        else {
+          u.NoSymbol
+        }
+      }
+      cls.info = u.ClassInfoType(cls.repr.parents, cls.repr.decls, assumedSelfSym)
       cls
     }
 
-    /** Normalises the parents and sets up value class machinery */
-    final def adjustParents(cls: Symbol, parents: List[Type]): List[Type] = {
-      val parentTypes = parents.map { tp =>
-        val tpe = tp.dealias
-        if (tpe.typeSymbolDirect === u.definitions.ObjectClass) u.definitions.AnyRefTpe
-        else tpe
-      }
+    /** sets up value class machinery */
+    final def processParents(cls: Symbol, parentTypes: List[Type]): parentTypes.type = {
       if (parentTypes.head.typeSymbolDirect === u.definitions.AnyValClass) {
         // TODO [tasty]: please reconsider if there is some shared optimised logic that can be triggered instead.
         withPhaseNoLater("extmethods") { ctx0 =>
@@ -412,7 +496,7 @@ trait ContextOps { self: TastyUniverse =>
       val moduleCls = sym.moduleClass
       val moduleClsFlags = FlagSets.withAccess(
         flags = FlagSets.Creation.ObjectClassDef,
-        inheritedAccess = sym.repr.originalFlagSet
+        inheritedAccess = sym.repr.tflags
       )
       val selfTpe = defn.SingleType(sym.owner.thisPrefix, sym)
       val ctor = newConstructor(moduleCls, selfTpe)
@@ -510,6 +594,35 @@ trait ContextOps { self: TastyUniverse =>
     def mode: TastyMode = EmptyTastyMode
     def owner: Symbol = topLevelClass.owner
 
+    private class TraceFrame(val id: Int, val next: TraceFrame) {
+
+      var nextChild: Int = 0
+
+      def show: String = {
+        val buf = mutable.ArrayDeque.empty[String]
+        var cur = this
+        while (cur.id != -1) {
+          buf.prepend(cur.id.toString)
+          cur = cur.next
+        }
+        buf.mkString("[", " ", ")")
+      }
+
+    }
+
+    private[this] var _trace: TraceFrame = new TraceFrame(id = -1, next = null)
+
+    private[ContextOps] def addFrame[T](op: String => T): T = {
+      val oldFrame  = _trace
+      val newFrame = new TraceFrame(id = oldFrame.nextChild, next = oldFrame)
+      _trace = newFrame
+      try op(newFrame.show)
+      finally {
+        _trace = oldFrame
+        _trace.nextChild += 1
+      }
+    }
+
     private[this] var mySymbolsToForceAnnots: mutable.LinkedHashSet[Symbol] = _
 
     private[ContextOps] def stageSymbolToForceAnnots(sym: Symbol): Unit = {
@@ -532,12 +645,19 @@ trait ContextOps { self: TastyUniverse =>
         val toForce = mySymbolsToForceAnnots.toList
         mySymbolsToForceAnnots.clear()
         for (sym <- toForce) {
-          log(s"!!! forcing annotations on ${showSym(sym)}")
-          analyseAnnotations(sym)
+          trace(traceForceAnnotations(sym)) {
+            analyseAnnotations(sym)
+          }
         }
         assert(mySymbolsToForceAnnots.isEmpty, "more symbols added while forcing")
       }
     }
+
+    private def traceForceAnnotations(sym: Symbol) = TraceInfo[Unit](
+      query = "forcing annotations of symbol",
+      qual = s"${showSym(sym)}",
+      res = _ => s"annotations were forced on ${showSym(sym)}"
+    )
 
     private[this] var myInlineDefs: mutable.Map[Symbol, mutable.ArrayBuffer[Symbol]] = null
     private[this] var myMacros: mutable.Map[Symbol, mutable.ArrayBuffer[Symbol]] = null
@@ -578,7 +698,7 @@ trait ContextOps { self: TastyUniverse =>
      * Reports illegal definitions:
      *   - trait constructors with parameters
      *
-     *  @param cls should be a symbol associated with a non-empty scope
+     *  @param cls should be a class symbol associated with a non-empty scope
      */
     private[ContextOps] def enterLatentDefs(cls: Symbol): Unit = {
